Add edit prop

+++ src/routes/songs/[id]/Song.svelte
@@ -5,9 +5,10 @@
 
 	interface Props {
 		songData: Observed<ReturnType<typeof getSong>>;
+		edit?: boolean;
 	}
 
-	let { songData }: Props = $props();
+	let { songData, edit }: Props = $props();
 
 	const url = $derived(`${base}/${songData.song.diskName}`);
 
@@ -48,65 +49,67 @@
 	Your browser does not support the audio element.
 </audio>
 
-<form bind:this={formElement} onsubmit={(e) => e.preventDefault()} class="mb-8">
-	<div class="grid grid-cols-2 gap-2">
-		<input type="number" id="start" disabled placeholder="0.0" step={sampleStep} />
-		<input type="number" id="end" disabled placeholder="0.0" step={sampleStep} />
-		<button
-			type="button"
-			class="btn mt-2 bg-blue-500 text-white hover:bg-blue-600"
-			onmousedown={() => {
-				if (formElement) {
-					const currentTime = getCurrentTime();
-					const input = formElement.querySelector<HTMLInputElement>("input[id='start']");
-					if (input) input.value = currentTime;
-				}
-			}}
-		>
-			Set loop start
-		</button>
+{#if edit}
+	<form bind:this={formElement} onsubmit={(e) => e.preventDefault()} class="mb-8">
+		<div class="grid grid-cols-2 gap-2">
+			<input type="number" id="start" disabled placeholder="0.0" step={sampleStep} />
+			<input type="number" id="end" disabled placeholder="0.0" step={sampleStep} />
+			<button
+				type="button"
+				class="btn mt-2 bg-blue-500 text-white hover:bg-blue-600"
+				onmousedown={() => {
+					if (formElement) {
+						const currentTime = getCurrentTime();
+						const input = formElement.querySelector<HTMLInputElement>("input[id='start']");
+						if (input) input.value = currentTime;
+					}
+				}}
+			>
+				Set loop start
+			</button>
 
-		<button
-			type="button"
-			class="btn mt-2 bg-blue-500 text-white hover:bg-blue-600"
-			onmousedown={() => {
-				if (formElement) {
-					const currentTime = getCurrentTime();
-					const input = formElement.querySelector<HTMLInputElement>("input[id='end']");
-					if (input) input.value = currentTime;
-				}
-			}}
-		>
-			Set loop end
-		</button>
-	</div>
+			<button
+				type="button"
+				class="btn mt-2 bg-blue-500 text-white hover:bg-blue-600"
+				onmousedown={() => {
+					if (formElement) {
+						const currentTime = getCurrentTime();
+						const input = formElement.querySelector<HTMLInputElement>("input[id='end']");
+						if (input) input.value = currentTime;
+					}
+				}}
+			>
+				Set loop end
+			</button>
+		</div>
 
-	<div>
-		<button
-			type="button"
-			class="btn mt-4 border-green-400 bg-green-500 text-white hover:bg-green-600"
-			onclick={async () => {
-				if (formElement) {
-					const startInput = formElement.querySelector<HTMLInputElement>("input[id='start']");
-					const endInput = formElement.querySelector<HTMLInputElement>("input[id='end']");
-					if (startInput && endInput) {
-						const start = parseFloat(startInput.value);
-						const end = parseFloat(endInput.value);
-						if (!isNaN(start) && !isNaN(end) && start < end) {
-							await addLoopToSong(songData.song.id, { start, end });
-							startInput.value = '';
-							endInput.value = '';
-						} else {
-							alert('Please enter valid start and end times. Start must be less than End.');
+		<div>
+			<button
+				type="button"
+				class="btn mt-4 border-green-400 bg-green-500 text-white hover:bg-green-600"
+				onclick={async () => {
+					if (formElement) {
+						const startInput = formElement.querySelector<HTMLInputElement>("input[id='start']");
+						const endInput = formElement.querySelector<HTMLInputElement>("input[id='end']");
+						if (startInput && endInput) {
+							const start = parseFloat(startInput.value);
+							const end = parseFloat(endInput.value);
+							if (!isNaN(start) && !isNaN(end) && start < end) {
+								await addLoopToSong(songData.song.id, { start, end });
+								startInput.value = '';
+								endInput.value = '';
+							} else {
+								alert('Please enter valid start and end times. Start must be less than End.');
+							}
 						}
 					}
-				}
-			}}
-		>
-			Add Loop
-		</button>
-	</div>
-</form>
+				}}
+			>
+				Add Loop
+			</button>
+		</div>
+	</form>
+{/if}
 
 {#each songData.songLoops as loop (loop.id)}
 	{@const isPlaying = a?.isLoopPlaying(loop.id)}
@@ -131,72 +134,75 @@
 			<span>Start: {loop.start.toFixed(1)} sec</span>
 			<span>End: {loop.end.toFixed(1)} sec</span>
 
-			<div class="flex gap-2">
-				<button
-					type="button"
-					class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
-					onclick={() => {
-						a?.playPreciseLoopEdges(loop.id, loop.start - 0.1, loop.end, {
-							rate: audioElement?.playbackRate ?? 1,
-							which: 'start',
-						});
-						updateLoop(loop.id, { start: loop.start - 0.1 });
-					}}
-				>
-					-
-				</button>
-				<button
-					type="button"
-					class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
-					onclick={() => {
-						a?.playPreciseLoopEdges(loop.id, loop.start + 0.1, loop.end, {
-							rate: audioElement?.playbackRate ?? 1,
-							which: 'start',
-						});
-						updateLoop(loop.id, { start: loop.start + 0.1 });
-					}}
-				>
-					+
-				</button>
-			</div>
+			{#if edit}
+				<div class="flex gap-2">
+					<button
+						type="button"
+						class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
+						onclick={() => {
+							a?.playPreciseLoopEdges(loop.id, loop.start - 0.1, loop.end, {
+								rate: audioElement?.playbackRate ?? 1,
+								which: 'start',
+							});
+							updateLoop(loop.id, { start: loop.start - 0.1 });
+						}}
+					>
+						-
+					</button>
+					<button
+						type="button"
+						class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
+						onclick={() => {
+							a?.playPreciseLoopEdges(loop.id, loop.start + 0.1, loop.end, {
+								rate: audioElement?.playbackRate ?? 1,
+								which: 'start',
+							});
+							updateLoop(loop.id, { start: loop.start + 0.1 });
+						}}
+					>
+						+
+					</button>
+				</div>
 
-			<div class="flex gap-2">
+				<div class="flex gap-2">
+					<button
+						type="button"
+						class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
+						onclick={() => {
+							a?.playPreciseLoopEdges(loop.id, loop.start, loop.end - 0.1, {
+								rate: audioElement?.playbackRate ?? 1,
+								which: 'end',
+							});
+							updateLoop(loop.id, { end: loop.end - 0.1 });
+						}}
+					>
+						-
+					</button>
+					<button
+						type="button"
+						class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
+						onclick={() => {
+							a?.playPreciseLoopEdges(loop.id, loop.start, loop.end + 0.1, {
+								rate: audioElement?.playbackRate ?? 1,
+								which: 'end',
+							});
+							updateLoop(loop.id, { end: loop.end + 0.1 });
+						}}
+					>
+						+
+					</button>
+				</div>
+
 				<button
 					type="button"
 					class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
 					onclick={() => {
-						a?.playPreciseLoopEdges(loop.id, loop.start, loop.end - 0.1, {
+						a?.playPreciseLoopEdges(loop.id, loop.start, loop.end, {
 							rate: audioElement?.playbackRate ?? 1,
-							which: 'end',
 						});
-						updateLoop(loop.id, { end: loop.end - 0.1 });
-					}}
+					}}>hear loop seam</button
 				>
-					-
-				</button>
-				<button
-					type="button"
-					class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
-					onclick={() => {
-						a?.playPreciseLoopEdges(loop.id, loop.start, loop.end + 0.1, {
-							rate: audioElement?.playbackRate ?? 1,
-							which: 'end',
-						});
-						updateLoop(loop.id, { end: loop.end + 0.1 });
-					}}
-				>
-					+
-				</button>
-			</div>
-			<button
-				type="button"
-				class="btn border-indigo-400 bg-indigo-500 text-white hover:bg-indigo-600"
-				onclick={() => {
-					a?.playPreciseLoopEdges(loop.id, loop.start, loop.end, {
-						rate: audioElement?.playbackRate ?? 1,
-					});
-				}}>hear loop seam</button
-			>
+			{/if}
 		</div>
 
 		<div class="mt-4 flex h-16 gap-2">
@@ -213,18 +219,20 @@
 				{isPlaying ? 'Stop' : 'Play'}
 			</button>
 
-			<button
-				type="button"
-				class="btn border-red-400 bg-red-500 text-white hover:bg-red-600"
-				onclick={() => {
-					a?.stopLoop(loop.id);
-					if (confirm(`Are you sure you want to delete loop "${loop.name}"?`)) {
-						deleteLoop(loop.id);
-					}
-				}}
-			>
-				Delete
-			</button>
+			{#if edit}
+				<button
+					type="button"
+					class="btn border-red-400 bg-red-500 text-white hover:bg-red-600"
+					onclick={() => {
+						a?.stopLoop(loop.id);
+						if (confirm(`Are you sure you want to delete loop "${loop.name}"?`)) {
+							deleteLoop(loop.id);
+						}
+					}}
+				>
+					Delete
+				</button>
+			{/if}
 		</div>
 	</div>
 {/each}
